requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
safetensors
opencv-python
fastapi
uvicorn
python-multipart
pyngrok
nest-asyncio
xformers
peft
bitsandbytes
pillow
numpy
scipy
insightface
onnxruntime-gpu
controlnet_aux
scikit-learn
pytest
httpx
# New dependencies for roadmap features
huggingface_hub>=0.20.0
websockets>=12.0
